+++ ChangeLog.md
@@ -3,6 +3,8 @@ XP AST ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Fixed `lang.ast.nodes.CatchStatement::children()` returning variable
+  (@thekid)
 * Fixed `lang.ast.nodes.ArrayLiteral::children()` returning NULL keys
   (@thekid)
 * Fixed `lang.ast.nodes.CaseLabel::children()` not returning case body